requirements.txt
# Core dependencies for local DeepSeek fine-tuning
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0
trl>=0.7.0
accelerate>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# For inference
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# For AI-generated training data
google-generativeai>=0.8.0

# For Tinker cloud training (install tinker-cookbook separately, see README)
# tinker>=0.6.1

# Optional: bitsandbytes for 4-bit quantization (may not work on macOS)
# pip install bitsandbytes  # Linux/CUDA only
